+++ src/main/java/daybyquest/dislike/application/DeletePostDislikeService.java
@@ -1,6 +1,6 @@
-package daybyquest.like.application;
+package daybyquest.dislike.application;
 
-import daybyquest.like.domain.PostDislikes;
+import daybyquest.dislike.domain.PostDislikes;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
